+++ docs/changelog.md
@@ -9,6 +9,7 @@
 - feature: aac/server_connections/sms.py - SMS Server Connection (new in 10.0.8)
 - fix: base/admin_ssh_keys.py - Ignore error when same ssh key exists under different name
 - pylint: change format() to f-strings
+- feature: base/tracing.py - Get tracing configuration (new in 10.0.8)
 
 ## 2025.3.28.0